# app/api/API_spec.yaml
openapi: 3.0.3
info:
  title: Digitization Toolkit Backend API
  description: |
    API for the Digitization Toolkit - an open-source, modular digitization platform
    for capturing, managing, and organizing archival materials using Raspberry Pi and 64MP ArduCam cameras.
    
    The API provides endpoints for:
    - User authentication and account management
    - Project and collection management
    - Record (document/archival object) metadata management
    - Image capture control and processing
    - Camera device discovery and calibration
    - System monitoring
  version: 1.0.0
  contact:
    name: UCSB AMP Lab
    url: https://github.com/UCSB-AMPLab/digitization-toolkit-backend
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: http://localhost:3000
    description: Production server (frontend proxy)
...
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      operationId: register_user
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '200':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '409':
          description: Username or email already exists
  /auth/login:
    post:
      summary: Login and get access token
      operationId: login_user
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT token for authentication
                  token_type:
                    type: string
                    example: bearer
        '401':
          description: Invalid credentials
        '403':
          description: User is inactive
  /auth/refresh:
    post:
      summary: Refresh access token
      operationId: refresh_token
      tags:
        - auth
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefresh'
        '401':
          description: Invalid or expired token
  /auth/password-reset:
    post:
      summary: Reset user password
      operationId: reset_password
      tags:
        - auth
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordReset'
      responses:
        '200':
          description: Password reset successful
        '401':
          description: Invalid credentials or token
  /auth/{user_id}:
    delete:
      summary: Delete user account
      operationId: delete_user
      tags:
        - auth
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User deleted successfully
        '403':
          description: Not authorized to delete this user
        '404':
          description: User not found

  # ============================================================================
  # Project Endpoints
  # ============================================================================
...
components:
  schemas:
...
    UserCreate:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          example: johndoe
        email:
          type: string
          format: email
          example: john@example.com
        password:
          type: string
          format: password
          example: securepassword123
    UserLogin:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    UserRead:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    PasswordReset:
      type: object
      required:
        - old_password
        - new_password
      properties:
        old_password:
          type: string
          format: password
        new_password:
          type: string
          format: password
    TokenRefresh:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: bearer

    # Project Schemas
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
...